Modernize CondCore/DBOutputService test stubs
