+++ CondCore/DBOutputService/test/stubs/LumiBasedUpdateAnalyzer.cc
@@ -1,12 +1,36 @@
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/Framework/interface/one/EDAnalyzer.h"
 #include "FWCore/ServiceRegistry/interface/Service.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "CondCore/DBOutputService/interface/OnlineDBOutputService.h"
 #include "CondFormats/BeamSpotObjects/interface/BeamSpotObjects.h"
 
-#include "LumiBasedUpdateAnalyzer.h"
 #include <iostream>
+#include <string>
+#include <chrono>
+
+namespace edm {
+  class ParameterSet;
+  class Event;
+  class EventSetup;
+}  // namespace edm
+
+// class declaration
+class LumiBasedUpdateAnalyzer : public edm::one::EDAnalyzer<> {
+public:
+  explicit LumiBasedUpdateAnalyzer(const edm::ParameterSet& iConfig);
+  virtual ~LumiBasedUpdateAnalyzer();
+  virtual void beginJob();
+  virtual void endJob();
+  virtual void beginLuminosityBlock(const edm::LuminosityBlock& lumiSeg, const edm::EventSetup& context);
+  virtual void endLuminosityBlock(const edm::LuminosityBlock& lumiSeg, const edm::EventSetup& iSetup);
+  virtual void analyze(const edm::Event& evt, const edm::EventSetup& evtSetup);
+
+private:
+  std::string m_record;
+  int m_ret;
+};
 
 LumiBasedUpdateAnalyzer::LumiBasedUpdateAnalyzer(const edm::ParameterSet& iConfig)
     : m_record(iConfig.getParameter<std::string>("record")), m_ret(-2) {}
