+++ CondCore/DBOutputService/test/stubs/MyDataAnalyzer.cc
@@ -1,27 +1,48 @@
-#include "FWCore/Framework/interface/EDAnalyzer.h"
+#include "FWCore/Framework/interface/one/EDAnalyzer.h"
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/Framework/interface/MakerMacros.h"
 #include "FWCore/ServiceRegistry/interface/Service.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "FWCore/Framework/interface/Event.h"
 #include "CondCore/DBOutputService/interface/PoolDBOutputService.h"
-//#include "CondCore/DBCommon/interface/LogDBEntry.h"
 #include "CondCore/CondDB/interface/Exception.h"
-//#include "CondCore/DBCommon/interface/DbTransaction.h"
 #include "CondFormats/Calibration/interface/Pedestals.h"
 
-#include "MyDataAnalyzer.h"
+#include <string>
 #include <cstdlib>
 #include <iostream>
+
+namespace edm {
+  class ParameterSet;
+  class Event;
+  class EventSetup;
+}  // namespace edm
+
+// class decleration
+class MyDataAnalyzer : public edm::one::EDAnalyzer<> {
+public:
+  explicit MyDataAnalyzer(const edm::ParameterSet& iConfig);
+  virtual ~MyDataAnalyzer();
+  virtual void analyze(const edm::Event& evt, const edm::EventSetup& evtSetup);
+  virtual void endJob();
+
+private:
+  std::string m_record;
+  bool m_LoggingOn;
+};
+
 MyDataAnalyzer::MyDataAnalyzer(const edm::ParameterSet& iConfig)
     : m_record(iConfig.getParameter<std::string>("record")), m_LoggingOn(false) {
   m_LoggingOn = iConfig.getUntrackedParameter<bool>("loggingOn");
   std::cout << "MyDataAnalyzer::MyDataAnalyzer" << std::endl;
 }
+
 MyDataAnalyzer::~MyDataAnalyzer() { std::cout << "MyDataAnalyzer::~MyDataAnalyzer" << std::endl; }
+
 void MyDataAnalyzer::analyze(const edm::Event& evt, const edm::EventSetup& evtSetup) {
   //
 }
+
 void MyDataAnalyzer::endJob() {
   std::cout << "MyDataAnalyzer::endJob " << std::endl;
   edm::Service<cond::service::PoolDBOutputService> mydbservice;
@@ -31,20 +52,20 @@ void MyDataAnalyzer::endJob() {
   }
   try {
     std::string tag = mydbservice->tag(m_record);
-    Pedestals* myped = new Pedestals;
+    Pedestals myped;
     if (mydbservice->isNewTagRequest(m_record)) {
       for (int ichannel = 1; ichannel <= 5; ++ichannel) {
         Pedestals::Item item;
         item.m_mean = 1.11 * ichannel;
         item.m_variance = 1.12 * ichannel;
-        myped->m_pedestals.push_back(item);
+        myped.m_pedestals.push_back(item);
       }
       //create
       cond::Time_t firstTillTime = mydbservice->endOfTime();
       cond::Time_t firstSinceTime = mydbservice->beginOfTime();
       std::cout << "firstSinceTime is begin of time " << firstSinceTime << std::endl;
       std::cout << "firstTillTime is end of time " << firstTillTime << std::endl;
-      mydbservice->writeOne(myped, firstSinceTime, m_record);
+      mydbservice->writeOneIOV(myped, firstSinceTime, m_record);
     } else {
       //append
       std::cout << "appending payload" << std::endl;
@@ -52,11 +73,11 @@ void MyDataAnalyzer::endJob() {
         Pedestals::Item item;
         item.m_mean = 0.15 * ichannel;
         item.m_variance = 0.32 * ichannel;
-        myped->m_pedestals.push_back(item);
+        myped.m_pedestals.push_back(item);
       }
       cond::Time_t thisPayload_valid_since = 5;
       std::cout << "appeding since time " << thisPayload_valid_since << std::endl;
-      mydbservice->writeOne(myped, thisPayload_valid_since, m_record);
+      mydbservice->writeOneIOV(myped, thisPayload_valid_since, m_record);
       std::cout << "done" << std::endl;
     }
     //example for log reading
@@ -97,10 +118,7 @@ void MyDataAnalyzer::endJob() {
     //std::cout<<er.what()<<std::endl;
   } catch (const cms::Exception& er) {
     throw cms::Exception("DBOutputServiceUnitTestFailure", "failed MyDataAnalyzer", er);
-  } /*catch(const std::exception& er){
-    std::cout<<"caught std::exception "<<er.what()<<std::endl;
-  }catch(...){
-    std::cout<<"Unknown error"<<std::endl;
-    }*/
+  }
 }
+
 DEFINE_FWK_MODULE(MyDataAnalyzer);
