+++ CondCore/DBOutputService/test/stubs/writeBlobComplex.cc
@@ -3,10 +3,28 @@
 #include "FWCore/Framework/interface/Event.h"
 #include "CondCore/DBOutputService/interface/PoolDBOutputService.h"
 #include "CondFormats/Calibration/interface/BlobComplex.h"
-
-#include "writeBlobComplex.h"
+#include "FWCore/Framework/interface/one/EDAnalyzer.h"
 
 #include <iostream>
+#include <string>
+
+namespace edm {
+  class ParameterSet;
+  class Event;
+  class EventSetup;
+}  // namespace edm
+
+// class decleration
+class writeBlobComplex : public edm::one::EDAnalyzer<> {
+public:
+  explicit writeBlobComplex(const edm::ParameterSet& iConfig);
+  ~writeBlobComplex();
+  virtual void analyze(const edm::Event&, const edm::EventSetup&);
+  virtual void endJob() {}
+
+private:
+  std::string m_RecordName;
+};
 
 writeBlobComplex::writeBlobComplex(const edm::ParameterSet& iConfig) : m_RecordName("BlobComplexRcd") {}
 
@@ -20,11 +38,11 @@ void writeBlobComplex::analyze(const edm::Event& evt, const edm::EventSetup& evt
     return;
   }
   try {
-    BlobComplex* me = new BlobComplex;
+    BlobComplex me;
     unsigned int serial = 123;
-    me->fill(serial);
+    me.fill(serial);
     std::cout << "writeBlobComplex::about to write " << std::endl;
-    mydbservice->writeOne(me, mydbservice->currentTime(), m_RecordName);
+    mydbservice->writeOneIOV(me, mydbservice->currentTime(), m_RecordName);
   } catch (const std::exception& er) {
     std::cout << "caught std::exception " << er.what() << std::endl;
     throw er;
