+++ CondCore/DBOutputService/test/stubs/writeInt.cc
@@ -1,17 +1,17 @@
-#include "FWCore/Framework/interface/EDAnalyzer.h"
-#include <string>
+#include "FWCore/Framework/interface/one/EDAnalyzer.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "FWCore/ServiceRegistry/interface/Service.h"
 #include "CondCore/DBOutputService/interface/PoolDBOutputService.h"
 
 #include <vector>
 #include <sstream>
+#include <string>
 
 #include <boost/serialization/vector.hpp>
 
 typedef std::vector<int> Payload;
 
-class writeInt : public edm::EDAnalyzer {
+class writeInt : public edm::one::EDAnalyzer<> {
 public:
   explicit writeInt(const edm::ParameterSet& iConfig);
   virtual void analyze(const edm::Event&, const edm::EventSetup&) {}
@@ -25,7 +25,7 @@ private:
 void writeInt::endJob() {
   edm::Service<cond::service::PoolDBOutputService> outdb;
 
-  outdb->writeOne(new std::vector<int>(1, me), me, cont);
+  outdb->writeOneIOV(std::vector<int>(1, me), me, cont);
 }
 
 writeInt::writeInt(const edm::ParameterSet& iConfig) : cont("oneInt"), me(iConfig.getParameter<int>("Number")) {}
